pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.3.1",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]
